admin role & some auth methods via gate

+++ app/Http/Controllers/PostController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\StorePost;
 use App\Models\BlogPost;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {/**
@@ -31,6 +32,7 @@ class PostController extends Controller
      */
     public function create()
     {
+//        $this->authorize('posts.create');
         return view('posts.create');
     }
 
@@ -42,6 +44,7 @@ class PostController extends Controller
      */
     public function store(StorePost $request)
     {
+
         $validated = $request->validated();
         $post = BlogPost::create($validated);
 
@@ -71,6 +74,12 @@ class PostController extends Controller
      */
     public function edit($id)
     {
+        $post = BlogPost::findOrFail($id);
+        $this->authorize($post);
+
+//        if( Gate::denies('update-post', $post)){
+//            abort(403, 'Yetkiniz Yok!');
+//        }
         return view('posts.edit', ['post' => BlogPost::findOrFail($id)]);
     }
 
@@ -83,7 +92,14 @@ class PostController extends Controller
      */
     public function update(StorePost $request, $id)
     {
+
+
         $post = BlogPost::findOrFail($id);
+        $this->authorize( $post);
+
+//        if( Gate::denies('update-post', $post)){
+//            abort(403, 'Yetkiniz Yok!');
+//        }
         $validated = $request->validated();
         $post->fill($validated);
         $post->save();
@@ -102,6 +118,11 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = BlogPost::findOrFail($id);
+
+        $this->authorize($post);
+//        if( Gate::denies('delete-post', $post)){
+//            abort(403, 'Silmeye yetkin yok');
+//        }
         $post->delete();
 
         session()->flash('status', 'Blog post was deleted!');
